chore: Update eslint configuration and add organization parameter
This commit updates the eslint configuration by removing the "plugin:github/recommended" rule and adding the "organization" parameter to the `getInputs` function in the `index.ts` file. The "plugin:github/recommended" rule was removed because it is no longer needed, and the "organization" parameter was added to allow the user to specify the organization when retrieving the list of users. This change improves the flexibility and customization of the code.